backend/requirements.txt
# Django and Django REST framework
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Database
# For SQLite (default)
# For PostgreSQL (uncomment if needed)
# psycopg2-binary==2.9.10

# PDF processing
PyPDF2==3.0.1

# Image processing
Pillow==10.3.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Development tools
black==24.3.0  # Code formatting
isort==5.13.2  # Import sorting
flake8==7.0.0  # Linting

# For deployment
gunicorn==23.0.0 